+++ lib/core/catcher.dart
@@ -13,10 +13,10 @@ import 'package:catcher/model/report.dart';
 import 'package:catcher/model/report_handler.dart';
 import 'package:catcher/model/report_mode.dart';
 import 'package:catcher/utils/catcher_error_widget.dart';
+import 'package:catcher/utils/catcher_logger.dart';
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:logging/logging.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 class Catcher with ReportModeAction {
@@ -44,12 +44,13 @@ class Catcher with ReportModeAction {
   /// Should catcher run WidgetsFlutterBinding.ensureInitialized() during initialization.
   final bool ensureInitialized;
 
-  final Logger _logger = Logger("Catcher");
   late CatcherOptions _currentConfig;
+  late CatcherLogger _logger;
+  late CatcherScreenshotManager screenshotManager;
   final Map<String, dynamic> _deviceParameters = <String, dynamic>{};
   final Map<String, dynamic> _applicationParameters = <String, dynamic>{};
   final List<Report> _cachedReports = [];
-  final CatcherScreenshotManager screenshotManager = CatcherScreenshotManager();
+  final Map<DateTime, String> _reportsOcurrenceMap = {};
   LocalizationOptions? _localizationOptions;
 
   /// Instance of navigator key
@@ -75,8 +76,8 @@ class Catcher with ReportModeAction {
   void _configure(GlobalKey<NavigatorState>? navigatorKey) {
     _instance = this;
     _configureNavigatorKey(navigatorKey);
-    _configureLogger();
     _setupCurrentConfig();
+    _configureLogger();
     _setupErrorHooks();
     _setupReportModeActionInReportMode();
     _setupScreenshotManager();
@@ -105,7 +106,6 @@ class Catcher with ReportModeAction {
     switch (ApplicationProfileManager.getApplicationProfile()) {
       case ApplicationProfile.release:
         {
-          _logger.fine("Using release config");
           if (releaseConfig != null) {
             _currentConfig = releaseConfig!;
           } else {
@@ -115,7 +115,6 @@ class Catcher with ReportModeAction {
         }
       case ApplicationProfile.debug:
         {
-          _logger.fine("Using debug config");
           if (debugConfig != null) {
             _currentConfig = debugConfig!;
           } else {
@@ -125,7 +124,6 @@ class Catcher with ReportModeAction {
         }
       case ApplicationProfile.profile:
         {
-          _logger.fine("Using profile config");
           if (profileConfig != null) {
             _currentConfig = profileConfig!;
           } else {
@@ -222,16 +220,18 @@ class Catcher with ReportModeAction {
   }
 
   void _configureLogger() {
-    if (enableLogger) {
-      Logger.root.level = Level.ALL;
-      Logger.root.onRecord.listen(
-        (LogRecord rec) {
-          // ignore: avoid_print
-          print(
-              '[${rec.time} | ${rec.loggerName} | ${rec.level.name}] ${rec.message}');
-        },
-      );
+    if (_currentConfig.logger != null) {
+      _logger = _currentConfig.logger!;
+    } else {
+      _logger = CatcherLogger();
     }
+    if (enableLogger) {
+      _logger.setup();
+    }
+
+    _currentConfig.handlers.forEach((handler) {
+      handler.logger = _logger;
+    });
   }
 
   void _loadDeviceInfo() {
@@ -471,6 +471,7 @@ class Catcher with ReportModeAction {
 
   ///Setup screenshot manager's screenshots path.
   void _setupScreenshotManager() {
+    screenshotManager = CatcherScreenshotManager(_logger);
     final String screenshotsPath = _currentConfig.screenshotsPath;
     if (!ApplicationProfileManager.isWeb() && screenshotsPath.isEmpty) {
       _logger.warning("Screenshots path is empty. Screenshots won't work.");
@@ -505,6 +506,8 @@ class Catcher with ReportModeAction {
       _setupLocalization();
     }
 
+    _cleanPastReportsOccurences();
+
     File? screenshot;
     if (!ApplicationProfileManager.isWeb()) {
       screenshot = await screenshotManager.captureAndSave();
@@ -521,6 +524,12 @@ class Catcher with ReportModeAction {
       _getPlatformType(),
       screenshot,
     );
+
+    if (_isReportInReportsOccurencesMap(report)) {
+      _logger.fine(
+          "Error: '$error' has been skipped to due to duplication occurence within ${_currentConfig.reportOccurrenceTimeout} ms.");
+      return;
+    }
 
     if (_currentConfig.filterFunction != null &&
         _currentConfig.filterFunction!(report) == false) {
@@ -541,6 +550,8 @@ class Catcher with ReportModeAction {
           "$reportMode in not supported for ${describeEnum(report.platformType)} platform");
       return;
     }
+
+    _addReportInReportsOccurencesMap(report);
 
     if (reportMode.isContextRequired()) {
       if (_isContextValid()) {
@@ -622,7 +633,7 @@ class Catcher with ReportModeAction {
       _logger.warning(
           "Error occurred in ${reportHandler.toString()}: ${handlerError.toString()}");
     }).then((result) {
-      _logger.info("Report result: $result");
+      _logger.info("${report.runtimeType} result: $result");
       if (!result) {
         _logger.warning("${reportHandler.toString()} failed to report error");
       } else {
@@ -693,6 +704,7 @@ class Catcher with ReportModeAction {
     };
   }
 
+  ///Get platform type based on device.
   PlatformType _getPlatformType() {
     if (ApplicationProfileManager.isWeb()) {
       return PlatformType.web;
@@ -714,6 +726,34 @@ class Catcher with ReportModeAction {
     }
 
     return PlatformType.unknown;
+  }
+
+  ///Clean report ocucrences from the past.
+  void _cleanPastReportsOccurences() {
+    int occurenceTimeout = _currentConfig.reportOccurrenceTimeout;
+    DateTime nowDateTime = DateTime.now();
+    _reportsOcurrenceMap.removeWhere((key, value) {
+      DateTime occurenceWithTimeout =
+          key.add(Duration(milliseconds: occurenceTimeout));
+      return nowDateTime.isAfter(occurenceWithTimeout);
+    });
+  }
+
+  ///Check whether reports occurence map contains given report.
+  bool _isReportInReportsOccurencesMap(Report report) {
+    if (report.error != null) {
+      return _reportsOcurrenceMap.containsValue(report.error.toString());
+    } else {
+      return false;
+    }
+  }
+
+  ///Add report in reports occurences map. Report will be added only when
+  ///error is not null and report occurence timeout is greater than 0.
+  void _addReportInReportsOccurencesMap(Report report) {
+    if (report.error != null && _currentConfig.reportOccurrenceTimeout > 0) {
+      _reportsOcurrenceMap[DateTime.now()] = report.error.toString();
+    }
   }
 
   ///Get current Catcher instance.