+++ lib/handlers/http_handler.dart
@@ -7,11 +7,9 @@ import 'package:catcher/model/report_handler.dart';
 import 'package:catcher/utils/catcher_utils.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
-import 'package:logging/logging.dart';
 
 class HttpHandler extends ReportHandler {
   final Dio _dio = Dio();
-  final Logger _logger = Logger("HttpHandler");
 
   final HttpRequestType requestType;
   final Uri endpointUri;
@@ -96,7 +94,7 @@ class HttpHandler extends ReportHandler {
 
   void _printLog(String log) {
     if (printLogs) {
-      _logger.info(log);
+      logger.info(log);
     }
   }