+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.6.7
+* Fixed typos by Morgan (https://github.com/rmtmckenzie) and Nico Mexis (https://github.com/ThexXTURBOXx)
+* Updated dependencies by Nico Mexis (https://github.com/ThexXTURBOXx)
+* Added `reportOccurrenceTimeout` in CatcherOptions to prevent reporting duplicated errors.
+* Added new handler: SnackbarHandler which can be used to show snackbar message.
+* Reworked logger in Catcher. Now Catcher uses only one class to handle log printing. Default logger can be overridden with custom logger.
+* Added `logger` parameter in CatcherOptions to provide custom logger to Catcher.
+* General logging/reporting fixes by naurmi (https://github.com/narumishi)
+* Added missing CMakeLists.txt for Linux.
+
 ## 0.6.6
 * Added copyWith method in LocalizationOptions.
 * Updated chinese translations in LocalizationOptions.